Adds check if Provider is initialized
If I use <Consumer /> outside of a wrapping <Provider /> atm, I get a non-helpful error like:

````
react-waterfall.js:1 Uncaught TypeError: Cannot read property 'state' of undefined
    at value (react-waterfall.js:1)
````

The proposed change checks for existance of the state-object and notifies otherwise. A possible point of debate could be the handling after the error. For now, it simple returns nothing.